Stock Watch
Version 0.4 - In Dev

This is an Android App that  displays stock prices with 15 mins delay. It is based on info from yahooapi.
This current version shows the history of the stock for up to a month. ofer time that would be up to 10 years.
THe provided widget get updated automatically.